Move files back to their spots
